done hw3_ios
