--- Navigation/ProfileViewController.swift
@@ -1,14 +1,0 @@
-import UIKit
-
-class ProfileViewController: UIViewController {
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        self.view.backgroundColor = .systemBrown
-        self.title = "Страница профиля"
-    }
-    
-    @objc private func didTapButton(){
-        let vc = PostViewController()
-        self.navigationController?.pushViewController(vc, animated: true)
-    }
-}
